Move FakeCliMulti to tests/PHPUnit/Framework/Mock so it can be reused.
